Chore: Move migration step implementations into package
